% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wlmm.R
\name{calcLL}
\alias{calcLL}
\title{Calculate log likelihood for a given heritability}
\usage{
calcLL(hsq, Phi, y, X, p, pwt, R)
}
\arguments{
\item{hsq}{heritability}

\item{Phi}{relatedness matrix}

\item{y}{phenotypes}

\item{X}{covariate matrix}

\item{p}{first-order sampling probability for all individuals}

\item{pwt}{pairwise sampling probability for sampled individuals}

\item{R}{sampling indicator (1 if sampled and 0 otherwise) for all individualsg}
}
\value{
The log likelihood value, with the corresponding estimates
of \code{beta} and \code{sigmasq} included as attributes.
}
\description{
Calculate the log likelihood for a given value of the heritability, \code{hsq}.
}
\examples{
data(sim)
loglik <- calcLL(0.5, Phi, y, X, p, pwt, R)
many_loglik <- calcLL(seq(0, 1, by=0.1), Phi, y, X, p, pwt, R)
}
