% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmm_modified.R
\name{calcLL2}
\alias{calcLL2}
\title{Calculate log likelihood for a given heritability}
\usage{
calcLL2(hsq, Kva, y, X)
}
\arguments{
\item{hsq}{heritability}

\item{Kva}{eigenvalues of K (calculated by \code{\link[=eigen_rotation2]{eigen_rotation2()}})}

\item{y}{rotated phenotypes (calculated by \code{\link[=eigen_rotation2]{eigen_rotation2()}})}

\item{X}{rotated covariate matrix (calculated by \code{\link[=eigen_rotation2]{eigen_rotation2()}})}

\item{pwt}{sampling probabilities}

\item{R}{a vector of 0 and 1 indicates which individuals are selected for resequencing}
}
\value{
The log likelihood value, with the corresponding estimates
of \code{beta} and \code{sigmasq} included as attributes.
}
\description{
Calculate the log likelihood for a given value of the heritability, \code{hsq}.
}
\examples{
data(recla)
e <- eigen_rotation2(recla$kinship, recla$pheno[,1], recla$covar)
loglik <- calcLL2(0.5, e$Kva, e$y, e$X)
many_loglik <- calcLL2(seq(0, 1, by=0.1), e$Kva, e$y, e$X)
}
