% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmm_modified.R
\name{eigen_rotation2}
\alias{eigen_rotation2}
\title{eigen decomposition + rotation}
\usage{
eigen_rotation2(K, y, pwt, R, X = NULL)
}
\arguments{
\item{K}{Kinship matrix}

\item{y}{Phenotypes}

\item{pwt}{sampling probabilities}

\item{R}{a vector of 0 and 1 indicates which individuals are selected for resequencing}

\item{X}{Numeric matrix with covariates. If NULL, use a column
of 1's (for intercept).}
}
\value{
List containing \code{Kva}, \code{Kve_t} and rotated
\code{y} and \code{X}.
}
\description{
Do eigen decomposition of kinship matrix and rotate \code{X} and
\code{y} by that, i.e., pre-multiply by the transpose of the matrix
of eigenvectors. If \code{Kva} and \code{Kve_t} provided, just do
the "rotation".
}
\examples{
data(recla)
e <- eigen_rotation2(recla$kinship, recla$pheno[,1], recla$covar)
}
