% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wlmm.R
\name{fitLMM}
\alias{fitLMM}
\title{Fit a linear mixed model}
\usage{
fitLMM(
  Phi,
  y,
  X,
  p,
  pwt,
  R,
  check_boundary = TRUE,
  tol = 1e-04,
  compute_se = FALSE
)
}
\arguments{
\item{Phi}{relatedness matrix}

\item{y}{phenotypes}

\item{X}{covariate matrix}

\item{p}{first-order sampling probability for all individuals}

\item{pwt}{pairwise sampling probability for sampled individuals}

\item{R}{sampling indicator (1 if sampled and 0 otherwise) for all individuals}

\item{check_boundary}{If TRUE, explicitly check log likelihood at 0 and 1.}

\item{tol}{Tolerance for convergence}

\item{compute_se}{= if TRUE, return the standard error of the \code{hsq}
estimate using the Fisher Information matrix of the MLE estimate. The
standard error will be in an \code{attr} of  \code{hsq} in the output.
Currently requires \code{use_cpp = FALSE}, and so if \code{compute_se=TRUE}
we take \code{use_cpp=FALSE}.}
}
\value{
List containing estimates of \code{beta}, \code{sigmasq},
\code{hsq}, \code{sigmasq_g}, and \code{sigmasq_e}, as well as the log
likelihood (\code{loglik}). If \code{compute_se=TRUE}, the output also
contains \code{hsq_se}.
}
\description{
Fit a linear mixed model of the form y = Xb + e where e follows a
multivariate normal distribution with mean 0 and variance matrix
\verb{sigmasq_g K + sigmasq_e I}, where \code{K} is a known kniship
matrix and \code{I} is the identity matrix.
}
\examples{
data(sim)
result <- fitLMM(Phi, y, X, p, pwt, R)

# also compute SE
wSE <- fitLMM(Phi, y, X, p, pwt, R, compute_se = TRUE, use_cpp=FALSE)
c(hsq=wSE$hsq, SE=wSE$hsq_se)

}
