% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wlmm.R
\name{getMLsoln}
\alias{getMLsoln}
\title{Get MLEs for coefficients and variance}
\usage{
getMLsoln(hsq, Phi, y, X, p, pwt, R)
}
\arguments{
\item{hsq}{heritability}

\item{Phi}{relatedness matrix}

\item{y}{phenotypes}

\item{X}{covariate matrix}

\item{p}{first-order sampling probability for all individuals}

\item{pwt}{pairwise sampling probability for sampled individuals}

\item{R}{sampling indicator (1 if sampled and 0 otherwise) for all individuals}
}
\value{
list containing \code{beta} and \code{sigmasq}, with residual
sum of squares as attributes.
}
\description{
For a fixed value for \code{hsq}, the heritability, calculate the
corresponding maximum likelihood estimates of \code{beta} and
\code{sigmasq}, with the latter being the total variance,
\code{sigmasq_g + sigmasq_e}.
}
\examples{
data(sim)
ml <- getMLsoln(0.5, Phi, y, X, p, pwt, R)
}
